requirements.txt
streamlit
python-dotenv
pydantic
sqlalchemy
alembic
chromadb
llama-index
openai
pypdf
tqdm
tiktoken
sentence-transformers
google-genai
requests
